// lib/meta-client/schema.ts
export interface components {
  schemas: {
// ...
    SchedulableFeatureViewEventKindEnum: "ingestion" | "profiling"
    /** @enum {string} */
    SchedulableEventKindEnum: "ingestion" | "profiling" | "store_profiling"
    /** @enum {string} */
    SchedulableStoreEventKindEnum: "store_profiling"
// ...
    EventParams: {
      partition?: components["schemas"]["PartitionCoord"]
    }
// ...
    ScheduleId: number
    FeatureViewSchedule: {
      schedule_id: components["schemas"]["ScheduleId"]
      feature_view_name: components["schemas"]["FeatureViewName"]
      event_kind: components["schemas"]["SchedulableFeatureViewEventKindEnum"]
      cron_expression?: string
      schedule_enabled: boolean
      params?: components["schemas"]["EventParams"]
      /** Format: date-time */
      created_at: string
      /** Format: date-time */
      updated_at: string
    }
    FeatureViewScheduleCreate: {
      feature_view_name: components["schemas"]["FeatureViewName"]
      event_kind: components["schemas"]["SchedulableFeatureViewEventKindEnum"]
      cron_expression?: string
      schedule_enabled?: boolean
      params?: components["schemas"]["EventParams"]
    }
    FeatureViewSchedulePatch: {
      cron_expression?: string
      schedule_enabled?: boolean
      params?: components["schemas"]["EventParams"]
    }
    FeatureViewScheduleList: {
      items: components["schemas"]["FeatureViewSchedule"][]
      pagination: components["schemas"]["Pagination"]
    }
    StoreEventSchedule: {
      schedule_id: components["schemas"]["ScheduleId"]
      store_name: components["schemas"]["StoreName"]
      store_kind: components["schemas"]["StoreKind"]
      entity_name: components["schemas"]["EntityCoordinate"]
      event_kind: components["schemas"]["SchedulableStoreEventKindEnum"]
      cron_expression?: string
      schedule_enabled: boolean
      /** Format: date-time */
      created_at: string
      /** Format: date-time */
      updated_at: string
    }
    StoreEventScheduleCreate: {
      store_name: components["schemas"]["StoreName"]
      store_kind: components["schemas"]["StoreKind"]
      entity_name: components["schemas"]["EntityCoordinate"]
      event_kind: components["schemas"]["SchedulableStoreEventKindEnum"]
      cron_expression?: string
      schedule_enabled?: boolean
    }
    StoreEventSchedulePatch: {
      cron_expression?: string
      schedule_enabled?: boolean
    }
    StoreEventScheduleList: {
      items: components["schemas"]["StoreEventSchedule"][]
      pagination: components["schemas"]["Pagination"]
    }
// ...
  }
// ...
}